feat(GI-01): types partages, TCB 256B canonique, crate SSR ExoPhoenix
- libs/exo_types/ : 12 fichiers sources (no_std, subtle dep)
  Types: PhysAddr, VirtAddr, ObjectId, CapToken, IpcMessage,
  FixedString, IoVec, PollFd, EpollEventAbi, ExoError, ZERO_BLOB_ID_4K
  Corrections: CORR-05,06,07,17,30,31,45,52 + SRV-02 + HASH-01

- kernel/src/scheduler/core/task.rs : TCB 256B align(64) remplace ancien
  kstack_ptr@[8], cr3_phys@[56], fpu_state_ptr@[232] (offsets hardcodes)
  sched_state:AtomicU64 encode etat|signal|flags (CORR-01)
  pid:ProcessId champ direct@[92], Priority+SchedPolicy types newtype
  ThreadId passe de u32 a u64, assertions offset_of! compilees

- libs/exo-phoenix-ssr/ : crate SSR ExoPhoenix
  SSR_BASE=0x0100_0000, SSR_SIZE=64 KiB, MAX_CORES=256 (CORR-02)
  freeze_ack_offset() et pmc_snapshot_offset() const fn

- Migration 31 fichiers kernel vers nouvelle API TCB
  scheduler/ process/ ipc/ syscall/ tous mis a jour

- cargo check: 0 erreur

docs/avancement/GI-01_Types_TCB_SSR_COMPTE_RENDU.md: compte rendu complet

+++ kernel/src/ipc/endpoint/connection.rs
@@ -80,7 +80,7 @@ impl HandshakeMsg {
             _pad:    0,
             cookie:  cookie.get(),
             ep_id:   ep_id.get(),
-            tid:     tid.0,
+            tid:     tid.0 as u32,
             _pad2:   0,
         }
     }
@@ -93,7 +93,7 @@ impl HandshakeMsg {
             _pad:    0,
             cookie:  cookie.get(),
             ep_id:   ep_id.get(),
-            tid:     server_tid.0,
+            tid:     server_tid.0 as u32,
             _pad2:   0,
         }
     }
